Added sample Tech Webs
Tech Web 1 is a more structured Web with 4 main "branches" or themed
paths that the player can take, amassing skill levels that they have to
spend wisely on increasingly expensive abilities. They can also
hybridize between 2 adjacent skills for some combination of skill paths.

Tech Web 2 is a more organic and free-flowing shaped one, where each
path only has a few key "Nodes" in it and each node has a series of
child or leaf skills that are directly derived from it and perhaps
modify the aforementioned node skill. This skill web is not as direct
and doesn't have as much room for various kinds of skills and abilities,
but allows for highly individual/modified skills and encourages
hybridization as well, with a hybrid "final tier" ability inbetween each
of the 4 main "paths".