Working collision.
We need to do something w/ old width and height.  They caused an offset
which explains the weird hacky code we had to write in places to get map
collision to work.  In order to make map collision perfect (ie not get
stuck at corners, etc.) we need to reflect these changes in it.